+++ docs/conf.py
@@ -9,9 +9,9 @@ import sys
 
 try:
     import nengo
-    import guzzle_sphinx_theme
+    import nengo_sphinx_theme
 except ImportError:
-    print("To build the documentation, nengo and guzzle_sphinx_theme must be "
+    print("To build the documentation, nengo and nengo_sphinx_theme must be "
           "installed in the current environment. Please install these and "
           "their requirements first. A virtualenv is recommended!")
     sys.exit(1)
@@ -40,14 +40,14 @@ extensions = [
     'sphinx.ext.mathjax',
     'sphinx.ext.todo',
     'sphinx.ext.viewcode',
-    'guzzle_sphinx_theme',
+    'nengo_sphinx_theme.ext.versions',
     'nbsphinx',
     'numpydoc',
 ]
 
 # -- sphinx.ext.autodoc
 autoclass_content = 'both'  # class and __init__ docstrings are concatenated
-autodoc_default_flags = ['members']
+autodoc_default_options = {"members": None}
 autodoc_member_order = 'bysource'  # default is alphabetical
 
 # -- sphinx.ext.intersphinx
@@ -60,6 +60,7 @@ intersphinx_mapping = {
 
 # -- sphinx.ext.todo
 todo_include_todos = True
+
 # -- numpydoc config
 numpydoc_show_class_members = False
 
@@ -88,18 +89,13 @@ pygments_style = "sphinx"
 templates_path = ["_templates"]
 html_static_path = ["_static"]
 
-html_theme_path = guzzle_sphinx_theme.html_theme_path()
-html_theme = "guzzle_sphinx_theme"
-
-html_theme_options = {
-    "project_nav_name": "Nengo core %s" % (version,),
-    "base_url": "https://www.nengo.ai/nengo",
-}
+html_theme = "nengo_sphinx_theme"
 
 html_title = "Nengo core {0} docs".format(release)
 htmlhelp_basename = 'Nengo core'
 html_last_updated_fmt = ''  # Suppress 'Last updated on:' timestamp
 html_show_sphinx = False
+html_sidebars = {"**": ["sidebar.html"]}
 
 # -- Options for LaTeX output -------------------------------------------------